requirements.txt
fastapi
uvicorn[standard]
httpx
google-cloud-bigquery
